Hands_on_Lab/Day_1/Task_4/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
groq==0.11.0
pydantic==2.9.2
python-dotenv==1.0.1